% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapDelta.R
\name{bootstrapDelta}
\alias{bootstrapDelta}
\title{Estimating standard errors for ATEs given non-random attrition}
\usage{
bootstrapDelta(regressionFormula, instrumentFormula, data,
  p_W_Formula = p_W_Formula, p_W_Method = p_W_Method,
  PiFormula = PiFormula, PiMethod = PiMethod, nBoots = 1000,
  quantiles = c(0.05, 0.95), effectType = "ATT")
}
\arguments{
\item{regressionFormula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{regressionFormula}
is the model which will be used to estimate the ATE after accounting for non-random
attrition. Formula must be of the form Y ~ D + ..., where Y is the outcome, D is the treatment,
and ... represents any set of other covariates.}

\item{instrumentFormula}{An object of class \code{formula} (or one that can be coerced to
that class); a symbolic description of the model to be fitted.  \code{instrumentFormula}
is the model which estimates the weights used in the model which estimates the ATE.
Formula must be of the form ~ Z1 + ..., where Z1 represents some instrumental variable,
and ... represents any other desired instrumental variables.}

\item{data}{A data frame which contains all variables to be used in both 
\code{regressionFormula} and \code{instrumentFormula}.}

\item{nBoots}{Numeric value defining the number of bootstrap samples; the default number
of bootstrap replications is 1,000.}

\item{weightMethod}{An optional string indicating the estimation method to be used for
generating the weights.  The default method is generalized linear model ("glm").}
}
\value{
A list of three vectors for mean, median, and standard errors of the ATE.
 \item{MeanEst}{A numeric vector of mean values of the ATE for treatment 
 and control group.}
 \item{MedianEst}{A numeric vector of the median of the ATE for treatment 
 and control group.}
 \item{SE}{A numeric vector of standard errors of the ATE for treatment 
 and control group.}
 \item{Quantiles}{A matrix containing estimates that correspond to the quantile (row)
 and coefficient (column). By defualt it returns the 5\% and 95\% quantiles.}
 \item{Matrix}{A matrix containing the bootstrapped coefficient estimates, where each
 row corresponds to a covaraite, and each column corresponds to a bootstrap iteration.}
}
\description{
\code{bootstrapDelta} estimates standard errors of the average treatment effect (ATE) 
under conditions of non-random attrition.
}
\details{
The function estimates standard errors for the average treatment effect (ATE)
after accounting for non-random attrition. After drawing random samples with 
replacement from the provided data (by default, 1000 samples), the function 
estimates with the bootstrapped sample the mean value, median, and standard errors 
based on the bootstrapped replications. The function, thus, provides a measure of 
accuracy to the point estimates for the ATE.
}
\references{
Huber, Martin (2012): "Identification of Average Treatment Effects in 
Social Experiments Under Alternative Forms of Attrition.", Journal of 
Educational and Behavioral Statistics, vol. 37, no. 3, 443-474.
}
\author{
Ryden Butler, David Miller, Jonas Markgraf, and Hyunjoo Oh
}
