% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cluster_CP.R
\name{check_cluster_CP}
\alias{check_cluster_CP}
\title{Check if there is a cluster in segmentation result}
\usage{
check_cluster_CP(Series_df, Name_series, CP)
}
\arguments{
\item{Series_df}{A data frame containing data for a main-nearby pair. It contains
at least two columns: the first labeled \code{Date} with dates in "YYYY-MM-DD" format
and the second being one of the six difference series, including: GPS - ERA, GPS - GPS',
GPS - ERA', ERA - ERA', GPS' - ERA', and GPS' - ERA, with names \code{GE, GGp,
GEp, EEp, GpEp,} and \code{GpE}, respectively.}

\item{Name_series}{The name of the column in \code{Series_df} to check.}

\item{CP}{A vector of dates in Date format: "%Y-%m-%d" representing
change-points in the series to check.}
}
\value{
Vector of break points in a cluster in Date format: "\%Y-\%m-\%d"
}
\description{
This function determine if any clusters of change-points exist in a station
based on segmentation results. A cluster is defined as a group of change-points
within 80 days.
}
