% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_XY_df.R
\name{construct_XY_df}
\alias{construct_XY_df}
\title{Create a data frame containing X and Y (named "signal") for regression}
\usage{
construct_XY_df(data, name_series, break_ind = NULL, Fourier = TRUE)
}
\arguments{
\item{data}{A data frame containing at least 2 columns named "Date" and
"Name_series". The "Date" column must be in Date format with the format "%Y-%m-%d",
and the "Name_series" column must be numeric (can contain NA values).}

\item{name_series}{The name of the column in "data" that we want to test,
as a character string.}

\item{break_ind}{The index of the break if the offset is included in the model,
as an integer.}

\item{Fourier}{A logical value indicating whether to include Fourier elements
in the model (TRUE or FALSE).}
}
\value{
A data frame used for testing that includes a column of the signal (Y)
and columns of X: 8 Fourier components (if selected), a global column, and a
jump column (if break_ind is not NULL).
}
\description{
Create a data frame containing X and Y (named "signal") for regression
}
\keyword{internal}
