% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_begin_end_date.R
\name{get_min_max_date}
\alias{get_min_max_date}
\title{find the begin and end of data frame without NA}
\usage{
get_min_max_date(data, column_name)
}
\arguments{
\item{data}{A data frame containing at least 2 columns named "Date" and
"Name_series". The "Date" column must be in Date format with the format "%Y-%m-%d",
and the "Name_series" column must be numeric (can contain NA values).}

\item{column_name}{The name of the column in "data" that we want to check,
as a character string.}
}
\value{
Vector of Date which are the beginning and the end of time series
with available data in column_name
}
\description{
find the begin and end of data frame without NA
}
\keyword{internal}
