% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLS.R
\name{GLS}
\alias{GLS}
\title{Fit GLS (known variance-covariance matrix)}
\usage{
GLS(phi, theta, var.t, df_XY)
}
\arguments{
\item{phi}{A numeric value indicating the parameter of the AR component.}

\item{theta}{A numeric value indicating the parameter of the MA component.}

\item{var.t}{A variance matrix computed by the moving window.}

\item{df_XY}{A data frame resulted from construct_XY_df that includes a column
of the signal (Y) and columns of X: 8 Fourier components (if selected), a global
mean column, and a jump column (if break_ind is not NULL).}
}
\value{
A list containing the estimates and their standard errors, t-values,
 and p-values.
}
\description{
Fit GLS (known variance-covariance matrix)
}
\keyword{internal}
