% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictive_rule}
\alias{predictive_rule}
\title{Apply the predictive rule to the test result}
\usage{
predictive_rule(test_result, predictive_models)
}
\arguments{
\item{test_result}{A data frame with dimension (n*6) where  where n is the
number of nearby stations. The 6 columns correspond to the 6 difference series.}

\item{predictive_models}{A list of 24 Random Forest pre-trained and given in
"inst/extdata.Rf.RDS"}
}
\value{
A dataframe of 3 columns the  candidate configuration with the highest
frequency after running 24 models, the probability of the candidate, and the
final configuration. Each row corresponds to a nearby station.
}
\description{
Apply the predictive rule to the test result
}
\keyword{internal}
