% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_std.R
\name{Sliding_std}
\alias{Sliding_std}
\title{Compute the sliding window standard deviation by ScaleTau estimator}
\usage{
Sliding_std(Y, name.var, length.wind = 60)
}
\arguments{
\item{Y}{A data frame contains at least 2 columns named Date and
name.var. The Date column must be in Date type on this format: "%Y-%m-%d"),
and the Name_series column is numeric (can be NA).}

\item{name.var}{The name of the column in "Y" that we want to test,
as a character string.}

\item{length.wind}{An integer value specifying the length of the window used
to compute the sliding standard deviation. Default is 60.}
}
\value{
A vector of values of the sliding standard deviation computed by the
scaleTau estimator.
}
\description{
Compute the sliding window standard deviation by ScaleTau estimator
}
\keyword{internal}
