% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_CP.R
\name{Test_CP}
\alias{Test_CP}
\title{Test Jump Significance in Difference Series}
\usage{
Test_CP(
  Series_df,
  Name_series,
  CP,
  limit = NULL,
  tol = 0.01,
  noise_model,
  length_win = 60,
  name_case = NULL
)
}
\arguments{
\item{Series_df}{A data frame containing data for a main-nearby pair. It contains
at least two columns: the first labeled \code{Date} with dates in "YYYY-MM-DD" format
and the second being one of the six difference series, including: GPS - ERA, GPS - GPS',
GPS - ERA', ERA - ERA', GPS' - ERA', and GPS' - ERA, with names \code{GE, GGp, GEp, EEp,
GpEp,} and \code{GpE}, respectively.}

\item{Name_series}{A string representing the name of the column in \code{Series_df} to check.}

\item{CP}{A date in Date format: "%Y-%m-%d" representing a change-point
in the main station.}

\item{limit}{An integer specifying the number of points in the segments
before and after the change-point used for testing. If NULL (default), the segments
before and after each tested change-point are defined by the nearest change-points.}

\item{tol}{A numeric value indicating the tolerance of the parameter to stop
the iterative FGLS.}

\item{noise_model}{A vector of 3 integer values indicating the order of the ARMA(p,0,q) model, where
p and q can be 0 or 1.}

\item{length_win}{An integer value specifying the length of the window used
to compute the moving variance.}

\item{name_case}{A string value indicating the name of the file to save the detailed results.
By default, the detailed results are not saved.}
}
\value{
A list of 2 elements:

1. A vector of dates that have available data used in the test.

2. A summary data frame of the test results with dimensions (l x 4), where l
is the number of variables in the model, and the 4 columns correspond to the
estimates, their standard errors, t-values, and p-values. By default, 10 variables
are included in the output, such as the mean, periodic bias (modeled by 4 coupled
sine and cosine Fourier series), and the offset.
}
\description{
Tests the significance of the jump at breakpoints in a series using Feasible
Generalized Least Squares.
}
