% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribution_CP.R
\name{Attribution_CP}
\alias{Attribution_CP}
\title{Predict the Origin of Change-Points}
\usage{
Attribution_CP(
  dataset,
  main_cp,
  nearby_cp,
  noise_model_fix = NULL,
  nearby_weight = NULL,
  limit_2side = 100
)
}
\arguments{
\item{dataset}{A list where:

  - Each element represents a main-nearby station pair, named after the nearby station.

  - Each element is a data frame with 7 columns:
    1. Date: In format "YYYY-MM-DD"
    2-7. GE, GGp, GEp, EEp, GpEp, GpE: Numeric, representing six difference series.

 The difference series are:
 GPS-ERA, GPS-GPS', GPS-ERA', ERA-ERA', GPS'-ERA', GPS'-ERA, respectively.}

\item{main_cp}{A vector of change-points in the main station (in Date format: "\%Y-\%m-\%d").}

\item{nearby_cp}{A list of vectors:

  - Each vector contains change-points for a nearby station (format: "YYYY-MM-DD").

  - Each vector named after the corresponding station.}

\item{noise_model_fix}{Can be:

  - A string specifying one noise model for all series.

  - A data frame (6 rows x n columns) specifying models for each series (row) and nearby station (column).

  - NULL to determine noise models using NoiseModel_Id.R.
  Valid models: 'AR(1)', 'MA(1)', 'ARMA(1,1)', 'White'.}

\item{nearby_weight}{A numeric vector where:

  - Length equals the number of nearby stations

  - Each element is the weight for a nearby station

  - If NULL, all stations are assigned equal weight}

\item{limit_2side}{An integer that:

  - Specifies the number of points used before and after each changepoint for testing

  - If NULL (default), all points are used

Please verify the format of input in the tests/testthat/test-attribution.R
CHECK HOW TO MAKE IT LINE BY LINE}
}
\value{
A list with four components:

\describe{
  \item{noise_model}{
A data frame (6 rows x n columns) where:

 * Rows represent the six difference series

 * Columns represent nearby stations

 * Each cell contains a string indicating the noise model (e.g., 'AR(1)',
MA(1)') of the according series }

  \item{test_results}{A list of data frames, where each data frame contains the test
results for each changepoint and each nearby station. The columns are the names of the
nearby stations, and the rows are the six difference series.}
  \item{prediction_results}{A list of data frames containing the predicted
configuration for each changepoint and nearby station. Each data frame is named
after the corresponding nearby station and contains three columns: the
candidate configuration with the highest frequency after running 24 models,
the probability of the candidate, and the final configuration. Each row
corresponds to a nearby station.}
  \item{aggregated_results}{ A data frame of the aggregated prediction results
for each changepoint, based on the weights specified in nearby_weight. This data
frame is presented in a similar format to the prediction_results data frame for
a single case}
}
}
\description{
This function predicts the origin of change-points in a station, which detected
on the difference series GPS - reference (ERA). The ideas of this method is to use
nearby station, which give GPS' and ERA' series. User need to compute the
six difference series  (GPS-ERA, GPS-GPS', GPS-ERA', ERA-ERA', GPS'-ERA', GPS'-ERA)
for each main-nearby station pair. This function tests the significance of changes
in mean in each of the six difference series for each main-nearby station pair,
predicts configurations from the test results, and aggregates them from all nearby stations
for each change-points.
}
\details{
The function operates in four main steps:

\strong{1. Model Identification} (optional): Uses the `NoiseModel_Id.R` function to
to identify the noise model for each difference series.

\strong{2. Significance Test}: Uses the `Test_CP.R` function to test the significance
of changes in mean in each difference series.

\strong{3. Configuration Prediction}:

   - Uses the `Prediction_CP.R` function to predict the configuration of each
triplet (main, changepoint, nearby) based on their test result of the six difference series.

   - Uses 24 pre-trained predictive models from `Predictiverule.RDS`.

\strong{4. Aggregation}: Aggregates the prediction results for each change-point
from all nearby stations.

Additionally, the function performs two preliminary checks:

\strong{1. Cluster Check}:

   - Looks for clusters of change-points (within 80 days) in main and nearby stations.

   - If a cluster exists, the function stops.

\strong{2. Coincident Check}:
\itemize{
  \item Identifies "coincident" change-points (within 10 days) between main and nearby stations.
  \item Coincident points in nearby stations are not used for selection of segment for significance test.
  \item Data between main change-points and coincident points is removed from four non-collocated series.
}
}
