% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cluster_CP.R
\name{check_cluster_CP}
\alias{check_cluster_CP}
\title{Check if there is cluster(s) in the segmentation results}
\usage{
check_cluster_CP(Series_df, Name_series, CP, threshold = 79)
}
\arguments{
\item{Series_df}{a data frame containing data for a main-nearby pair.
It contains at least two columns: the first column,labeled \code{date}, contains the dates in Date format (\%Y-\%m-\%d)
and the second contains one of the six difference series, including GPS - ERA, GPS - GPS',
GPS - ERA', ERA - ERA', GPS' - ERA', and GPS' - ERA, with names \code{GE, GGp,
GEp, EEp, GpEp,} and \code{GpE}, respectively}

\item{Name_series}{the name of the column in \code{Series_df} to check}

\item{CP}{a vector of dates in Date format (\%Y-\%m-\%d) given the change-points in the series to check}

\item{threshold}{a number indicating the maximum length of a cluster}
}
\value{
a vector of change-points which belongs to a cluster in Date format
}
\description{
This function determines if any clusters of change-points exist in the segmentation
results for a series. A cluster is defined as a group of change-points
separated at least by 80 days
}
