% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cluster_CP.R
\name{check_cluster_CP}
\alias{check_cluster_CP}
\title{Check if there is a cluster in segmentation result}
\usage{
check_cluster_CP(Series_df, Name_series, Break_points)
}
\arguments{
\item{Series_df}{A data frame contains at least 2 columns named Date and
Name_series. The Date column must be in Date type on this format: "\%Y-\%m-\%d"),
and the Name_series column is numeric (can be NA).}

\item{Name_series}{The name of the column in "data" that we want to test,
as a character string.}

\item{Break_points}{Vector of breakpoints in the test series in Date type
on this format: "\%Y-\%m-\%d".}
}
\value{
Vector of break points in a cluster in Date type
}
\description{
Check if there is a cluster in segmentation result
}
\keyword{internal}
