% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_similar_CP.R
\name{check_similar_CP}
\alias{check_similar_CP}
\title{To check for similar changes in nearby stations}
\usage{
check_similar_CP(Series_df, main_cp, nearby_cp_one, threshold)
}
\arguments{
\item{Series_df}{A data frame containing data for a main-nearby pair. It contains seven
columns: the first labeled \code{Date} with dates in "YYYY-MM-DD" format and the
remaining six columns containing numeric values representing six different
series: GPS - ERA, GPS - GPS', GPS - ERA', ERA - ERA', GPS' - ERA', and GPS' - ERA,
with names \code{GE, GGp, GEp, EEp, GpEp,} and \code{GpE}, respectively.}

\item{main_cp}{A vector of dates in Date format: "%Y-%m-%d" representing
change-points in the main station.}

\item{nearby_cp_one}{A vector of dates in Date format: "%Y-%m-%d" representing
change-points in the nearby station.}

\item{threshold}{An integer specifying the number of points between two change-points,
which defines if it is a "similar" change-point. Default is 10.}
}
\value{
A vector of dates in Date format: "%Y-%m-%d" representing "similar"
change-points in the nearby station to the change-points in the main station.
}
\description{
This function identifies "similar" change-points in a nearby station that occur
within a specified number of days of the change-points in the main station.
Data between change-points in the main station and "similar" change-points
in the nearby station is removed from the four non-collocated series for the
significance test of the change in mean.
}
