% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_similar_CP.R
\name{check_similar_CP}
\alias{check_similar_CP}
\title{To check is there similar changes in the nearby stations}
\usage{
check_similar_CP(Six_Series, List_break_main, List_break_nearby, threshold)
}
\arguments{
\item{Six_Series}{A data frame contains 6 time series differences (so 7 columns
in total) between a main station and a nearby station. The first column should
be the date (in Date type on this format: "\%Y-\%m-\%d"), and the next six columns
must be in this order:
G-E, G-G', G-E', E-E', G'-E', G'-E.}

\item{List_break_main}{A vector of main station breakpoints in Date type on this format:
"\%Y-\%m-\%d".}

\item{List_break_nearby}{A vector of main station breakpoints in Date type on this format:
"\%Y-\%m-\%d".}

\item{threshold}{Integer number is the threshold to define the number of
point between 2 breakpoints in a cluster}
}
\value{
list of main changes and their similar in the nearby in Date type
}
\description{
To check is there similar changes in the nearby stations
}
\keyword{internal}
