% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoiseModel_Id.R
\name{NoiseModel_Id}
\alias{NoiseModel_Id}
\title{Identify Time Series Noise Model}
\usage{
NoiseModel_Id(dataset, main_break, nearby_break)
}
\arguments{
\item{dataset}{A list of data frames, where each data frame contains 6 time series
differences (so 7 columns in total) between a main station and a nearby station. The
first column should be the date, and the next six columns must be in this order:
G-E, G-G', G-E', E-E', G'-E', G'-E. The name of each data frame must be the name
of the nearby station.}

\item{main_break}{A vector of main station breakpoints in Date type on this format:
"\%Y-\%m-\%d".}

\item{nearby_break}{A list of break point vectors for each nearby station, named
after the respective nearby station, in Date type on this format: "\%Y-\%m-\%d".}
}
\value{
A noise model data frame with dimensions (6 * n), where n is the
number of nearby stations. The 6 rows correspond to the 6 difference series.
The noise model for each series is given as a character string, such as "AR(1)",
"MA(1)", "ARMA(1,1)", or "White"
}
\description{
Identify Time Series Noise Model
}
