% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_CP.R
\name{Test_CP}
\alias{Test_CP}
\title{Test Jump Significance in Difference Series}
\usage{
Test_CP(
  Series_df,
  Name_series,
  CP,
  limit = NULL,
  tol = 0.01,
  noise_model,
  length_win = 60,
  name_case = NULL
)
}
\arguments{
\item{Series_df}{A data frame contains at least 2 columns named Date and
Name_series. The Date column must be in Date type on this format: "\%Y-\%m-\%d"),
and the Name_series column is numeric (can be NA).}

\item{Name_series}{The name of the column in "data" that we want to test,
as a character string.}

\item{CP}{A change point in Date type on this format: "\%Y-\%m-\%d"}

\item{limit}{An integer specifying the number of points to be used in the
significance testing before and after each breakpoint. This can be used to speed up
the testing process. Default is NULL, which means all points are used.}

\item{tol}{A numeric value indicating the tolerance of the parameter to stop
the iterative FGLS}

\item{noise_model}{A vector of 3 number indicating the ARMA(p,0,q) model, where
p and q can be 0 or 1}

\item{length_win}{An integer value specifying the length of the window used
to compute the sliding standard deviation.}

\item{name_case}{A logical value indicating whether to save the results or not.
By default, it does not save the results.}
}
\value{
A list of 2 components:
1. A vector of dates used in the test.
2. A data frame with dimensions (n x 4), where n is the number of variables
in the model, and the 4 columns correspond to the estimates, their standard errors,
t-values, and p-values.
}
\description{
Tests the significance of jumps at breakpoints in six difference series
between a main station and nearby stations.
}
